pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skillguard"
version = "0.1.0"
description = "VirusTotal for AI Agent Skills - Multi-engine security scanner for AI skills, MCP servers, and agentic tools"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "David Girard" }]
keywords = ["security", "ai", "skills", "mcp", "scanning", "prompt-injection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    # API
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-multipart>=0.0.18",
    # CLI
    "click>=8.1",
    "rich>=13.9",
    # Scanning engines
    "pyyaml>=6.0",
    # Data models
    "pydantic>=2.10",
    # Reporting
    "jinja2>=3.1",
    # Monitoring
    "watchdog>=6.0",
    # Utilities
    "httpx>=0.28",
]

[project.optional-dependencies]
yara = ["yara-python>=4.5"]
ml = [
    "onnxruntime>=1.20",
    "transformers>=4.47",
    "sentence-transformers>=3.3",
]
db = [
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.14",
    "asyncpg>=0.30",
]
semgrep = ["semgrep>=1.100"]
secrets = ["detect-secrets>=1.5"]
all = ["skillguard[yara,ml,db,semgrep,secrets]"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "httpx",
    "ruff>=0.8",
]

[project.scripts]
skillguard = "skillguard.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/skillguard"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
